fix: separate RAG examples into ingestion and app files
